changed js inserted button into static html markup - have javascript in separate imported file
