+++ site/layouts/partials/jumbotron.html
@@ -2,7 +2,7 @@
   <div class="mw7 center ph3">
     <div class="db mb3">
       <div class="mw7 relative bg-fix-primary mb3">
-        <h1 class="f2 f1-l b di lh-title mb3 white mw6 bg-primary">
+        <h1 class="f2 f1-l b di lh-title mb3 white mw6 bg-primary" id="entryTitle">
           {{.title}}
         </h1>
       </div>
