+++ site/layouts/post/single.html
@@ -1,7 +1,7 @@
 {{ partial "head" . }}
 {{ partial "nav" . }}
   <div class="mw6 center ph3 pv4">
-    <h1 class="f2 lh-title b mt0 mb3">{{.Title}}</h1>
+    <h1 class="f2 lh-title b mt0 mb3" id="entryTitle">{{.Title}}</h1>
     <div class="flex justify-between grey-3">
       <p class="mt0">{{ .Date.Format "Mon, Jan 2, 2006" }}</p>
       <p class="mt0">Read in {{ .ReadingTime }} minutes</p>
